move Destroyable utilities out of native_mate
