+++ filenames.gni
@@ -475,6 +475,8 @@ filenames = {
     "shell/common/crash_reporter/linux/crash_dump_handler.h",
     "shell/common/crash_reporter/win/crash_service_main.cc",
     "shell/common/crash_reporter/win/crash_service_main.h",
+    "shell/common/gin/destroyable.cc",
+    "shell/common/gin/destroyable.h",
     "shell/common/gin_converters/callback_converter_gin_adapter.h",
     "shell/common/gin_converters/file_dialog_converter.cc",
     "shell/common/gin_converters/file_dialog_converter.h",
